fastapi-app/requirements.txt
fastapi
uvicorn
celery[redis]
redis
ffmpeg-python
fastapi-cors
flower
